Optimize previous fix:
We only need to request the name from the name cache once. In case a name has
to be resolved another time while the response to our first request hasn't
arrived yet, do not request it again but just add another observer for that
agent id - it will be called as well once the request is finished.