STORM-2105: Switch to VersionHelper API to determine Windows version
While we could set major and minor operating system version with the data
returned by GetVersionEx / GetVersion, we set them explicitly as named function
may be altered or removed in the future. In that case, only the build number
would be affected.

Also removed the old way of determing the build number via shell version. This
was only needed for Windows 9x, which aren't supported anymore and can't even
run the viewer.