Set the avatar name expiration period via LL's method instead of applying a minimum expiration of 60 minutes
A minimum expiration of 60 minutes for an avatar name appears to be way too long
as the default expiration time returned by the region is 2 minutes. This means
people will see an outdated name 30 times longer than it should be.

Since this might expose an additional load to the region it can't handle due
to the previous less frequent name requests, the debug setting "FSLegacyNameCacheExpiration"
has been added to revert to our legacy behavior as instant "fix" in case this
will cause serious issues.