Removed the 'remove "Join" button from group invitations when joining costs a fee' functionality This reverts commit 85433378a363.
Removed, because it's primary functionality when it was added to the viewer was to protect new residents from clicking blindly on group invitations and spending money, but it was disabled by default anyway from the beginning.
New residents don't usually lurk into preferences too deeply anyway, and non-new residents usually know what they click on... plus there is an alert in the middle of the screen after trying to join a group with fee, too.
Second reason is the fact, that most probably it won't work with the new syswell thingy too well.