FIRE-6809: Quickly moving the bandwidth slider has no effect
Previous implementation was wrong and did not update the throttle to
whatever bandwidth setting was chosen within the last 0.25s of the bandwidth
being moved.

Adopt the Update class for the bandwidth from LLPanelPreferences into
llviewercontrol.cpp. It will make sure the throttle is only updated if there
hasn't been a change (e.g. slider being moved) for the last 0.5s. It will also
make sure that the last chosen value is used to adjust the throttle.