Fix LL impostor hacking: * Don't render shadows for impostors * Show muted avatars as impostors (fix if RenderAvatarMaxNonImpostors = 0 = no limit) * No detailed attachment updates for muted avatars if using no impostors * Properly calculate update interval for muted avatars if using no impostors
Changed LLVOAvatar::isImpostor() so it also returns if an avatar is an impostor
in case RenderAvatarMaxNonImpostors = 0 = LLVOAvatar::sUseImpostors = FALSE is set,
so we can get rid of some

"isImpostor() || ((LLVOAvatar::AV_DO_NOT_RENDER == getVisualMuteSettings() || isInMuteList()) && !needsImpostorUpdate())"

and replace it with

"isImpostor() && !needsImpostorUpdate()"

This additional check would be necessary to render an avatar as impostor in case
it should never be rendered or it is muted, but LLVOAvatar::sUseImpostors is FALSE.
The check for !needsImpostorUpdate() is still needed to fail the condition in case
an impostor needs an update. For impostor updates, LLPipeline::generateImpostor()
will temporarily set LLVOAvatar::sUseImpostors = FALSE, so the rendering pipeline
doesn't skip parts needed to create a silhouette of the avatar. In that case,
needsImpostorUpdate() will return TRUE.