FIRE-22742: Make English default language for installer
According to the NSIS manual (http://nsis.sourceforge.net/Docs/Chapter4.html#langs),
the language selection works as follows:

---
When the installer starts up it goes through these steps to select the interface language:

    Get user's default Windows UI language
    Find a perfect match for the language
    If there is no perfect match, find a primary language match
    If there is no match, use the first language defined in the script (make sure your first language is a common one like English)
    If the language variable $LANGUAGE has changed during .onInit, NSIS goes through steps 2 to 4 again.
---

Languages are defined via "LoadLanguageFile" command, that in our case are
inside the included lang_xx.nsi file. So make sure to load en-us first to
define it as default language being used for Windows installations that have
another language not part of the installer.