SL-10190: Add menu commands to force AV or exception in coroutine.
"Bad memory access" and "unhandled exception" are the two categories of error
that we expect might be different in a coroutine than in the viewer's main
fiber. Without this change, we've had no reliable way to force either of those
to occur.

This will require translation work for two new menu items.

# Conflicts:
#	indra/newview/skins/default/xui/en/menu_viewer.xml