SL-10415, SL-10434
Further tweaking of directional shadow sampling to balance between peter-panning and shadow acne.

Move stars to just this side of the sky some to reduce parallax.

Remove decodeNormF and uses of decode_normal in favor of unified use of getNorm
(try to help the Intel HD x000 compiler learn to link correctly).