SL-10486, SL-10501, SL-10487
Make moon texture from default daycycle the default moon texture id and use it for the default and blank assets in the picker.

Fix handling of MM_TEXTURE to correctly detect when texture units outside the range 0-3
are used with texture matrix ops to prevent mem overwrite bug and/or asserts.