SL-10896
Fix deferred water soften prog lightnorm getting stomped with rotated lightnorm,
because it's a water shader and needs water atmo uniform values,
but it's a deferred lighting shader that needs not-so-rotated lightnorms.

Make sunlight_color not get auto-updated so we can get the correct value from mSunDiffuse.

Remove mix to preserve bumps as it was washing out shadows.