SL-11216: Introduce generic LLStoreListener<T> to capture event data.
LLStoreListener is an adapter initialized with a reference to an LLEventPump
on which to listen, a reference to a variable into which to store received
data, and an optional llsd::drill() path to extract desired data from each
event received on the subject LLEventPump.

In effect, LLStoreListener is like a miniature LLEventAPI whose only operation
is to store to its destination variable.