DRTVWR-476: Validate LLEventPumpOrPumpName replyPump
passed to postAndSuspendsetup().

The requestPump is optional, and the function varies its behavior depending on
whether that parameter is empty or meaningful. But it unconditionally uses the
replyPump. Passing an empty LLEventPumpOrPumpName caused mysterious crashes.
Add llassert_always_msg() to make the coding error explicit in such a case.

Also streamline access to meaningful requestPump and replyPump by temporarily
caching the bound LLEventPump reference.